+++ src/crud_views/lib/table/attrs.py
@@ -76,6 +76,4 @@ class ColAttr(metaclass=ColAttrMeta):
     w100: ColumnAttrs = ColumnAttrs.td_class("cv-col-100")
 
     # extra
-    action: ColumnAttrs = ColumnAttrs.th_class("d-flex justify-content-end") | ColumnAttrs.td_class(
-        "d-flex justify-content-end"
-    )
+    action: ColumnAttrs = ColumnAttrs.th_class("cv-col-action") | ColumnAttrs.td_class("cv-col-action")